@@ -20,16 +20,19 @@ public class ShareFile {
     private String url;
     private Uri uri;
     private String type;
+    private String filename;
 
-    public ShareFile(String url, String type, ReactApplicationContext reactContext){
-        this(url, reactContext);
+    public ShareFile(String url, String type, String filename, ReactApplicationContext reactContext){
+        this(url, filename, reactContext);
         this.type = type;
+        this.filename = filename;
     }
 
-    public ShareFile(String url, ReactApplicationContext reactContext){
+    public ShareFile(String url, String filename, ReactApplicationContext reactContext){
         this.url = url;
         this.uri = Uri.parse(this.url);
         this.reactContext = reactContext;
+        this.filename = filename;
     }
     /**
      * Obtain mime type from URL
@@ -104,12 +107,13 @@ public class ShareFile {
 
         if(this.isBase64File()) {
             String encodedImg = this.uri.getSchemeSpecificPart().substring(this.uri.getSchemeSpecificPart().indexOf(";base64,") + 8);
+            String filename = this.filename != null ? this.filename : System.nanoTime() + "";
             try {
                 File dir = new File(Environment.getExternalStorageDirectory(), Environment.DIRECTORY_DOWNLOADS );
                 if (!dir.exists() && !dir.mkdirs()) {
                     throw new IOException("mkdirs failed on " + dir.getAbsolutePath());
                 }
-                File file = new File(dir, System.nanoTime() + "." + extension);
+                File file = new File(dir, filename + "." + extension);
                 final FileOutputStream fos = new FileOutputStream(file);
                 fos.write(Base64.decode(encodedImg, Base64.DEFAULT));
                 fos.flush();
