@@ -2,18 +2,17 @@ package com.example;
 
 import android.app.Application;
 
+import com.facebook.react.PackageList;
+import com.facebook.hermes.reactexecutor.HermesExecutorFactory;
+import com.facebook.react.bridge.JavaScriptExecutorFactory;
 import com.facebook.react.ReactApplication;
-import cl.json.RNSharePackage;
-import cl.json.ShareApplication;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
 
-import java.util.Arrays;
 import java.util.List;
 
-public class MainApplication extends Application implements ReactApplication, ShareApplication {
+public class MainApplication extends Application implements ReactApplication {
 
   private final ReactNativeHost mReactNativeHost = new ReactNativeHost(this) {
     @Override
@@ -23,10 +22,13 @@ public class MainApplication extends Application implements ReactApplication, Sh
 
     @Override
     protected List<ReactPackage> getPackages() {
-      return Arrays.<ReactPackage>asList(
-          new MainReactPackage(),
-          new RNSharePackage()
-      );
+
+      @SuppressWarnings("UnnecessaryLocalVariable")
+      List<ReactPackage> packages = new PackageList(this).getPackages();
+      // Packages that cannot be autolinked yet can be added manually here, for example:
+      //packages.add(new MyReactNativePackage());
+
+      return packages;
     }
 
     @Override
@@ -45,10 +47,4 @@ public class MainApplication extends Application implements ReactApplication, Sh
     super.onCreate();
     SoLoader.init(this, /* native exopackage */ false);
   }
-
-  @Override
-  public String getFileProviderAuthority() {
-        return BuildConfig.APPLICATION_ID + ".provider";
-  }
-
 }
